Add CommentForm component
Renders a form allowing the user to submit a new comment. A name and
message are both required.

Data is passed to the backend with a POST request using fetch. If the
request is successful, the new comment is added and displayed instantly.
Otherwise, relevant validation messages are displayed to the user.